Initial commit of Iceping project
Added the main source code of the Iceping tool.
Included all the necessary files and dependencies to build and run the project.
Created a README.md file to provide a brief overview of the project and its features.
Set up the project structure in a way that's easy to understand and maintain.